fix validateTime and updateSchema
